fix(dataflow): widen mesh_rows to uint16_t in block_mover_flow_executor (#25)
graph_.m_tiles is uint16_t (operand_flow_graph.hpp:364) but
block_mover_flow_executor.hpp:476 was assigning it to a local uint8_t
mesh_rows. This silently truncates for any mesh with >= 256 rows
and would break the `i + 1 < mesh_rows` bounds check on the next
line - a latent correctness bug.

Widening the local to uint16_t to match the source type fixes the
narrowing without changing observable behaviour for the typical
4x4 mesh.

Also clears the only library-side MSVC C4244 warning (issue #23
Phase 2a). The warning was being reported twice in the build log
because two TUs (test_c_stationary_matmul.cpp and
test_flow_graph_executors.cpp) include the header.

Verified locally on Linux gcc release:
  c_stationary_matmul_test: 14/14 cases,  24/24 assertions
  flow_graph_executor_test: 21/21 cases, 106/106 assertions

Refs #23

+++ include/sw/kpu/models/dataflow/block_mover_flow_executor.hpp
@@ -472,8 +472,12 @@ public:
         auto push_b = graph_.add_fire(Operation::PUSH_TO_L2, {b_l3}, {b_l2}, "push_B");
         graph_.add_edge(join_b, push_b);
 
-        // Forward B south (if not bottom row)
-        uint8_t mesh_rows = graph_.m_tiles;  // Assume square mesh or use m_tiles
+        // Forward B south (if not bottom row).
+        // graph_.m_tiles is uint16_t; using uint8_t here would silently
+        // wrap to 0 for any mesh >= 256 rows and break the bounds check,
+        // so match the source type. Also fixes MSVC C4244 narrowing
+        // warning that propagated to every TU including this header.
+        uint16_t mesh_rows = graph_.m_tiles;  // assumes square mesh
         if (i + 1 < mesh_rows) {
             auto b_south = tile_b(k, j, Location::L3, node_id + mesh_cols_);
             auto send_b = graph_.add_fire(Operation::SEND_SOUTH, {b_l3}, {b_south}, "send_B_south");
